Extract stats dashboard into its own module
Move the ~450-line HTML/CSS/Chart.js dashboard out of server.js into a
dedicated server/dashboard.js. The module is presentation-only: it takes a
plain data context and returns a string, with no dependency on live server
state or Supabase, which makes it independently testable.

- server.js now fetches the data (Supabase access stays with the server)
  and delegates rendering via renderDashboard(context); the two Supabase
  fetches run in parallel
- escapeHtml/jsonForScript move into the module alongside the markup they
  protect; the script-safe JSON embed still neutralizes hostile variant
  names in the chart data
- Drop dead code carried over from the original: variantRows and
  displayStats were computed but never inserted into the output
